requirements.txt
Cython == 0.29.32
Pillow == 7.1.2
PyYAML == 6.0
SoundFile == 0.10.3.post1
absl_py == 1.2.0
activations == 0.1.0
aidenva == 0.0.1
aml == 0.0.0
application == 1.0
aria2stub == 0.9.1
asposebarcode == 1.0.0
asposestorage == 1.0.2
automium-web == 0.1.1
bitarray == 2.6.0
bitcoin-toolkit == 0.0.2
boto3 == 1.24.46
botocore == 1.27.46
ccds-maskdetection == 0.1.0
ccds-maskdetection-version-01 == 0.2.0
ccds-maskdetection-version-02 == 0.2.0
cellml == 0.1.1
cgm-3d-cnn == 0.0.2
crystalml == 0.1.1
custodian-streetcred-id == 0.0.1
cvxpy == 1.2.1
data == 0.4
data-service == 0.0.35
datasets == 2.4.0
detr == 0.1.4
dl-timeseries == 0.0.14
dna-yolov5-torch == 1.0.1
dolg-tensorflow == 0.0.1
draupnir == 0.0.193
ecommerce-common-package == 1.1.8
editdistance == 0.5.3
einops == 0.4.1
examplejupyter == 1.0.0
fairseq == 0.12.2
fasttext == 0.9.2
flask-of-cinema == 1.0.0
ftfy == 6.0.3
gTTS == 2.2.4
gbrss == 0.0.0.2
googletrans == 4.0.0rc1
h5py == 3.1.0
hematopy == 0.0.1.dev6
hydra_core == 1.0.7
hyperband-multithreading == 0.0.1
indicnlp == 0.0.1
indras-net == 2.0.6
inflect == 2.1.0
intel-quantization == 1.1
intranet-invoices == 0.0.1
intranet-partner == 0.0.1
joblib == 1.1.0
kaldi-io == 0.9.4
kenlm == 0.0.0
kia-remarketing-supply == 1.0.0
librosa == 0.8.1
lightconv-layer == 0.0.1
lod-dedovschina-sber == 1.0.4
lupseat == 1.3.0
made-ai-dungeon == 0.0.4
marklib == 1.0.1
matplotlib == 3.2.2
mlzoo == 0.0.1
mysecondazibit == 0.0.1
nltk == 3.7
npy-append-array == 0.9.13
numpy == 1.21.6
octopytorch == 0.3
omegaconf == 2.0.6
opencv_contrib_python == 4.6.0.66
pandas == 1.3.5
prejudice-bias-hotel-reviews == 0.0.1
progressbar2 == 3.38.0
psutil == 5.4.8
py-gerrit == 1.0.2
pyarrow == 6.0.1
pycocoevalcap == 1.2
pycocotools == 2.0.4
pynotion == 0.0.2
pyskyscanner == 1.0.1
pythainlp == 3.0.8
pytorch_lightning == 1.7.0
randomshapelets == 0.0.9
realm-wishes == 0.0.2
regex == 2022.6.2
reluble == 0.1.0
requests == 2.23.0
riksdagenapi == 0.0.1
rmn == 3.0.4
roomai == 0.1.16
rouge_score == 0.1.2
sacrebleu == 2.2.0
sbirandm == 2.1.0
scipy == 1.7.3
sego-cli == 0.1.0.0
semantic-segmentation == 0.1.0
sentence-import-worker-nleaser == 0.4.0
setuptools == 57.4.0
sharingattention == 0.0.1
six == 1.15.0
sklearn == 0.0
soco-yolo == 0.0.9
spacy == 3.4.1
stockdl == 0.2.5
streamlit == 1.11.1
streetcred-id-service-clients == 1.0
sts-models-tj == 0.0.65
surfing-cli == 2.6.8
swarmer == 0.5.0
swiftcodegen == 0.3
tabulate == 0.8.10
tasks == 2.8.0
tensorboardX == 2.4.1
tf2-initializer == 0.6
timm == 0.6.7
todolist-v1 == 1.0.19
todolist-v2 == 1.0.21
todolisthub-v2 == 1.0.3
torch == 1.12.0+cu113
torch-soft == 0.1.1
torchaudio == 0.12.0+cu113
torchblocks == 1.0.0
torchvision == 0.13.0+cu113
tqdm == 4.64.0
tracko == 0.1.0
transferflow == 0.1.8
unidecode == 1.3.4
utils == 1.0.1
videodatavideodata == 1.0.0
vidmodex == 0.9a0
vino == 0.0.2
visunn == 0.1.1
visvmtagger == 1.0a0
wallstreetbets-sentiment-analyser == 0.0.3
wenlan-video == 1.0.0
wordcloud-create-worker-nleaser == 0.4.0
xpe-library == 0.0.3
yolov5-evaluator == 0.0.22
yolov5processor == 1.2
yrag == 0.1.5
zh-ito-yolov5 == 1.0.0.0